Bug with setting wrong paths for generator was fixed. Text fields for generator and validator arguments were modified. Importing java.io.* was added to GeneratorImplementation.java in FileSystemExample folder. Converting to long type was fixed in TokenInput class. CheckingListener was added in JavaClassChecker class (not used, with commented code).

+++ ProblemsCreator/Development/src/main/java/com/ran/development/check/TokenInput.java
@@ -48,7 +48,7 @@ public class TokenInput {
     
     public long nextLong() throws CheckResultException {
         try {
-            return Integer.parseInt(nextToken());
+            return Long.parseLong(nextToken());
         } catch (NumberFormatException exception) {
             throw new CheckResultException(failureResultInfo, "Long number was expected", exception);
         }
